Improvements: - nginx config parser now 2x time faster than before (thanks Paul McGuire for help) - nginx access log parser now faster than before
Fixes:
- small bugs discovered with Amazon linux, SLES and Fedora linux
- bug with ip addresses in net interfaces meta info

Closes: https://github.com/nginxinc/nginx-amplify-agent/issues/26, https://github.com/nginxinc/nginx-amplify-agent/issues/30

+++ test/base.py
@@ -129,9 +129,12 @@ class RealNginxTestCase(BaseTestCase):
         super(RealNginxTestCase, self).setup_method(method)
         self.second_started = False
         subp.call('service nginx start')
+        self.running = True
 
     def teardown_method(self, method):
-        subp.call('pgrep nginx |sudo xargs kill -SIGKILL')
+        if self.running:
+            subp.call('pgrep nginx |sudo xargs kill -SIGKILL')
+            self.running = False
         super(RealNginxTestCase, self).teardown_method(method)
 
     def reload_nginx(self):
